+++ pkg/crd/crds/apps-job.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,15 +26,13 @@ import (
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 
 	"github.com/arangodb/go-driver"
+
+	"github.com/arangodb/kube-arangodb/pkg/util"
 )
 
 const (
 	AppsJobVersion = driver.Version("1.0.1")
 )
-
-func init() {
-	mustLoadCRD(appsJobs, appsJobsSchemaRaw, &appsJobsCRD, &appsJobsCRDSchemas)
-}
 
 // Deprecated: use AppsJobWithOptions instead
 func AppsJob() *apiextensions.CustomResourceDefinition {
@@ -57,8 +55,8 @@ func AppsJobDefinitionWithOptions(opts ...func(*CRDOptions)) Definition {
 	}
 }
 
-var appsJobsCRD apiextensions.CustomResourceDefinition
-var appsJobsCRDSchemas crdSchemas
+var appsJobsCRD = util.NewYamlLoader[apiextensions.CustomResourceDefinition](appsJobs)
+var appsJobsCRDSchemas = util.NewYamlLoader[crdSchemas](appsJobsSchemaRaw)
 
 //go:embed apps-job.yaml
 var appsJobs []byte