+++ pkg/crd/crds/database-clustersynchronization.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,15 +26,13 @@ import (
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 
 	"github.com/arangodb/go-driver"
+
+	"github.com/arangodb/kube-arangodb/pkg/util"
 )
 
 const (
 	DatabaseClusterSynchronizationVersion = driver.Version("1.0.1")
 )
-
-func init() {
-	mustLoadCRD(databaseClusterSynchronization, databaseClusterSynchronizationSchemaRaw, &databaseClusterSynchronizationCRD, &databaseClusterSynchronizationCRDSchemas)
-}
 
 // Deprecated: use DatabaseClusterSynchronizationWithOptions instead
 func DatabaseClusterSynchronization() *apiextensions.CustomResourceDefinition {
@@ -57,8 +55,8 @@ func DatabaseClusterSynchronizationDefinitionWithOptions(opts ...func(*CRDOption
 	}
 }
 
-var databaseClusterSynchronizationCRD apiextensions.CustomResourceDefinition
-var databaseClusterSynchronizationCRDSchemas crdSchemas
+var databaseClusterSynchronizationCRD = util.NewYamlLoader[apiextensions.CustomResourceDefinition](databaseClusterSynchronization)
+var databaseClusterSynchronizationCRDSchemas = util.NewYamlLoader[crdSchemas](databaseClusterSynchronizationSchemaRaw)
 
 //go:embed database-clustersynchronization.yaml
 var databaseClusterSynchronization []byte