+++ pkg/crd/crds/database-member.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,15 +26,13 @@ import (
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 
 	"github.com/arangodb/go-driver"
+
+	"github.com/arangodb/kube-arangodb/pkg/util"
 )
 
 const (
 	DatabaseMemberVersion = driver.Version("1.0.1")
 )
-
-func init() {
-	mustLoadCRD(databaseMember, databaseMemberSchemaRaw, &databaseMemberCRD, &databaseMemberCRDSchemas)
-}
 
 // Deprecated: use DatabaseMemberWithOptions instead
 func DatabaseMember() *apiextensions.CustomResourceDefinition {
@@ -57,8 +55,8 @@ func DatabaseMemberDefinitionWithOptions(opts ...func(*CRDOptions)) Definition {
 	}
 }
 
-var databaseMemberCRD apiextensions.CustomResourceDefinition
-var databaseMemberCRDSchemas crdSchemas
+var databaseMemberCRD = util.NewYamlLoader[apiextensions.CustomResourceDefinition](databaseMember)
+var databaseMemberCRDSchemas = util.NewYamlLoader[crdSchemas](databaseMemberSchemaRaw)
 
 //go:embed database-member.yaml
 var databaseMember []byte