+++ pkg/crd/crds/ml-extension.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,15 +26,13 @@ import (
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 
 	"github.com/arangodb/go-driver"
+
+	"github.com/arangodb/kube-arangodb/pkg/util"
 )
 
 const (
 	MLExtensionVersion = driver.Version("1.0.0")
 )
-
-func init() {
-	mustLoadCRD(mlExtension, mlExtensionSchemaRaw, &mlExtensionCRD, &mlExtensionCRDSchemas)
-}
 
 func MLExtensionWithOptions(opts ...func(*CRDOptions)) *apiextensions.CustomResourceDefinition {
 	return getCRD(mlExtensionCRD, mlExtensionCRDSchemas, opts...)
@@ -47,8 +45,8 @@ func MLExtensionDefinitionWithOptions(opts ...func(*CRDOptions)) Definition {
 	}
 }
 
-var mlExtensionCRD apiextensions.CustomResourceDefinition
-var mlExtensionCRDSchemas crdSchemas
+var mlExtensionCRD = util.NewYamlLoader[apiextensions.CustomResourceDefinition](mlExtension)
+var mlExtensionCRDSchemas = util.NewYamlLoader[crdSchemas](mlExtensionSchemaRaw)
 
 //go:embed ml-extension.yaml
 var mlExtension []byte