+++ pkg/crd/crds/ml-job-cron.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,15 +26,13 @@ import (
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 
 	"github.com/arangodb/go-driver"
+
+	"github.com/arangodb/kube-arangodb/pkg/util"
 )
 
 const (
 	MLCronJobVersion = driver.Version("1.0.0")
 )
-
-func init() {
-	mustLoadCRD(mlCronJob, mlCronJobSchemaRaw, &mlCronJobCRD, &mlCronJobCRDSchemas)
-}
 
 func MLCronJobWithOptions(opts ...func(*CRDOptions)) *apiextensions.CustomResourceDefinition {
 	return getCRD(mlCronJobCRD, mlCronJobCRDSchemas, opts...)
@@ -47,8 +45,8 @@ func MLCronJobDefinitionWithOptions(opts ...func(*CRDOptions)) Definition {
 	}
 }
 
-var mlCronJobCRD apiextensions.CustomResourceDefinition
-var mlCronJobCRDSchemas crdSchemas
+var mlCronJobCRD = util.NewYamlLoader[apiextensions.CustomResourceDefinition](mlCronJob)
+var mlCronJobCRDSchemas = util.NewYamlLoader[crdSchemas](mlCronJobSchemaRaw)
 
 //go:embed ml-job-cron.yaml
 var mlCronJob []byte