+++ pkg/crd/crds/ml-storage.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,15 +26,13 @@ import (
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 
 	"github.com/arangodb/go-driver"
+
+	"github.com/arangodb/kube-arangodb/pkg/util"
 )
 
 const (
 	MLStorageVersion = driver.Version("1.0.0")
 )
-
-func init() {
-	mustLoadCRD(mlStorage, mlStorageSchemaRaw, &mlStorageCRD, &mlStorageCRDSchemas)
-}
 
 func MLStorageWithOptions(opts ...func(*CRDOptions)) *apiextensions.CustomResourceDefinition {
 	return getCRD(mlStorageCRD, mlStorageCRDSchemas, opts...)
@@ -47,8 +45,8 @@ func MLStorageDefinitionWithOptions(opts ...func(*CRDOptions)) Definition {
 	}
 }
 
-var mlStorageCRD apiextensions.CustomResourceDefinition
-var mlStorageCRDSchemas crdSchemas
+var mlStorageCRD = util.NewYamlLoader[apiextensions.CustomResourceDefinition](mlStorage)
+var mlStorageCRDSchemas = util.NewYamlLoader[crdSchemas](mlStorageSchemaRaw)
 
 //go:embed ml-storage.yaml
 var mlStorage []byte